+++ src/app/lib/util.ts
@@ -11,6 +11,7 @@ import {FormProperty} from '@lhncbc/ngx-schema-form';
 import {DateUtil} from './date-util';
 import {v4 as uuidv4} from 'uuid';
 import {fhirPrimitives} from "../fhir";
+declare var LForms: any;
 
 export type GuidingStep = 'home' | 'fl-editor' | 'item-editor';
 export enum FHIR_VERSIONS {
@@ -230,23 +231,32 @@ export class Util {
 
 
   /**
-   * Convert lforms units to equivalent FHIR extensions.
+   * Convert lforms units to equivalent FHIR extensions. For quantity type, all
+   * units are converted, and for decimal or integer, only the first unit is converted.
+   *
    * @param units - units in lforms format.
+   * @param dataType - 'quantity' || 'decimal' || 'integer'
    */
-  static convertUnitsToExtensions(units): any [] {
+  static convertUnitsToExtensions(units, dataType: string): any [] {
     if(!units) {
       return null;
     }
     const ret: any [] = [];
-    units.forEach((unit) => {
+    const unitUri = dataType === 'quantity' ?
+      'http://hl7.org/fhir/StructureDefinition/questionnaire-unitOption' :
+      'http://hl7.org/fhir/StructureDefinition/questionnaire-unit';
+    units.some((unit) => {
+      const display = LForms.ucumPkg.UcumLhcUtils.getInstance().validateUnitString(unit.unit)?.unit?.name || unit.unit;
       ret.push({
-        url: 'http://hl7.org/fhir/StructureDefinition/questionnaire-unit',
+        url: unitUri,
         valueCoding: {
-          code: unit,
+          code: unit.unit,
           system: 'http://unitsofmeasure.org',
-          display: unit
+          display: display
         }
       });
+      // For quantity convert all units. For decimal or integer pick the first one.
+      return (dataType !== 'quantity');
     });
     return ret;
   }